% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{address_locations}
\alias{address_locations}
\title{Data of random address locations in Amsterdam}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{id}{unique ID}
\item{geom}{geometry of the address location}
}
}
\source{
{Created Random data points in Amsterdam}
}
\usage{
address_locations
}
\description{
Contains 100 data points in Amsterdam
}
\examples{
data(address_locations)     # Lazy loading

}
\keyword{datasets}
