% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ams_Houses}
\alias{Ams_Houses}
\title{an sf data frame of houses in Amsterdam
Houses are the centroids of the neighborhoods in Amsterdam}
\format{
A data frame with 518 rows and 1 variables:
\describe{
\item{geom}{The geometry of the points of the houses}
}
}
\source{
{https://maps.amsterdam.nl/open_geodata/}
}
\usage{
Ams_Houses
}
\description{
Contains 518 points of houses in Amsterdam
}
\examples{
data(Ams_Houses)     # Lazy loading

}
\keyword{datasets}
