% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ams_Neighborhoods}
\alias{Ams_Neighborhoods}
\title{an sf data frame of neighborhoods in Amsterdam}
\format{
A data frame with 518 rows and 3 variables:
\describe{
\item{Buurtcode}{unique ID for each neighborhood}
\item{Buurt}{The name of each neighborhood}
\item{geom}{The geometry of the polygons of the neighborhoods}
}
}
\source{
{https://maps.amsterdam.nl/open_geodata/}
}
\usage{
Ams_Neighborhoods
}
\description{
Contains 518 polygons of neighborhoods in Amsterdam
}
\examples{
data(Ams_Neighborhoods)     # Lazy loading

}
\keyword{datasets}
