% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ams_Parks}
\alias{Ams_Parks}
\title{an sf data frame of parks in Amsterdam
Parks are multi polygons}
\format{
A data frame with 123 rows and 2 variables:
\describe{
\item{name}{The name of the park}
\item{geom}{The geometry of the points of the houses}
}
}
\source{
{https://maps.amsterdam.nl/open_geodata/}
}
\usage{
Ams_Parks
}
\description{
Contains 123 polygons of parks in Amsterdam
}
\examples{
data(Ams_Parks)     # Lazy loading

}
\keyword{datasets}
