% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bboxapp.R
\name{bboxapp}
\alias{bboxapp}
\title{Draw bounding box for the area of interest and get it in local environment}
\usage{
bboxapp()
}
\value{
once the user draw a rectangle, it return as bbox in the local environment
}
\description{
Draw bounding box for the area of interest and get it in local environment
}
