% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_calc.R
\name{calc_ndvi_new}
\alias{calc_ndvi_new}
\title{Creating average NDVI values per location}
\usage{
calc_ndvi_new(
  address_location,
  raster,
  buffer_distance = NULL,
  net = NULL,
  UID = NULL,
  address_calculation = TRUE,
  speed = NULL,
  time = NULL,
  engine = c("pc", "gee")
)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{raster}{raster file with NDVI values}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{net}{An optional sfnetwork object representing a road network, If missing the road network will be created.}

\item{UID}{A character string representing a unique identifier for each point of interest}

\item{address_calculation}{A logical, indicating whether to calculate the address location (if not a point) as the centroid of the polygon containing it (default is 'TRUE')}

\item{speed}{A numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{A numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{engine}{When the raster is missing, users can choose whether they want to use Google Earth engine \code{gee} or Planetary Computer \code{pc} to calculate the ndvi}
}
\value{
A \code{sf} dataframe with the mean ndvi, the geometry and the buffer that was used
}
\description{
Creating average NDVI values per location
}
