% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_coverage.R
\name{canopy_pct}
\alias{canopy_pct}
\title{Calculate the percentage of a canopy within a given buffer distance or location}
\usage{
canopy_pct(
  address_location,
  canopy_layer,
  buffer_distance = NULL,
  network_buffer = FALSE,
  network_file = NULL,
  epsg_code = NULL,
  folder_path_network = NULL,
  UID = NULL,
  address_location_neighborhood = FALSE,
  speed = NULL,
  time = NULL,
  city = NULL
)
}
\arguments{
\item{address_location}{A \code{\link[sf]{sf}} object representing the location of interest, the location should be in projected coordinates.}

\item{canopy_layer}{A \code{\link[sf]{sf}} that represents a canopy, the layer should be in projected coordinates}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{network_buffer}{Optional; a  logical, the default is an euclidean buffer, when TRUE, a network buffer will be used.}

\item{network_file}{Optional; a \code{\link[sfnetworks]{sfnetwork}} object representing a road network, If missing the road network will be created.}

\item{epsg_code}{Optional; a epsg code to get a Projected CRS in the final output, If missing, the default is \code{3395}}

\item{folder_path_network}{Optional; a Folder path to where the retrieved network should be saved continuously. Must not include a filename extension (e.g. '.shp', '.gpkg').}

\item{UID}{Optional; a character string representing a unique identifier for each point of interest}

\item{address_location_neighborhood}{Optional; a logical, indicating whether to calculate with an address point or a neighbourhood. default is \code{FALSE}}

\item{time}{Optional; a  numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{city}{Optional; a when using a network buffer, you can add a city where your address points are to speed up the process}
}
\value{
The percentage of the canopy within a given buffer or isochrone around a set of locations is printed.
}
\description{
Calculate the percentage of a canopy within a given buffer distance or location
}
\examples{
#' # Read the data
df <- sf::st_read('path/to/data')

# read the canopy layer
canopy_layer <- sf::st_read('path/to/canopy')

# calculate the canopy percentage for the POIs with default settings
canopy_pct(address_locations=df, canopy=canopy_layer, buffer_distance=300)

# calculate canopy percentage for the neighborhood, using the network as buffer
canopy_pct(address_locations=df, buffer_distance=300, network_buffer=TRUE)


}
