% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_coverage.R
\name{canopy_perc}
\alias{canopy_perc}
\title{Calculate the percentage of a canopy within a given buffer distance or location}
\usage{
canopy_perc(
  address_location,
  canopy_layer,
  buffer_distance = NULL,
  network_buffer = FALSE,
  network_file = NULL,
  UID = NULL,
  address_calculation = TRUE,
  speed = NULL,
  time = NULL,
  city = NULL
)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{canopy_layer}{A canopy layer that represents a canopy, the layer should be in projected coordinates}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{network_buffer}{A logical, the default is an euclidean buffer, when TRUE, a network buffer will be used.}

\item{network_file}{An optional sfnetwork object representing a road network, If missing the road network will be created.}

\item{UID}{A character string representing a unique identifier for each point of interest}

\item{address_calculation}{A logical, indicating whether to calculate the address location (if not a point) as the centroid of the polygon containing it (default is 'TRUE')}

\item{speed}{A numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{A numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{city}{When using a network buffer, you can add a city where your address points are to speed up the process}
}
\value{
The percentage of the canopy within a given buffer or isochrone around a set of locations is printed.
}
\description{
Calculate the percentage of a canopy within a given buffer distance or location
}
