% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{greenspace_access}
\alias{greenspace_access}
\title{Distance to the nearest fake entries of greenspace}
\usage{
greenspace_access(
  address_location,
  greenspace = NULL,
  buffer_distance = NULL,
  network_file = NULL,
  epsg_code = NULL,
  download_dir = tempdir(),
  euclidean = TRUE,
  pseudo_entrance = FALSE,
  entrances_within_buffer = TRUE,
  speed = NULL,
  time = NULL,
  city = NULL,
  UID = NULL
)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{greenspace}{A spatial object representing greenspace}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{network_file}{An optional sfnetwork object representing a road network, If missing the road network will be created.}

\item{epsg_code}{A espg code to get a Projected CRS in the final output, If missing, the default is \code{3395}}

\item{download_dir}{A directory to download the network file, the default will be \code{tempdir()}.}

\item{euclidean}{Whether the distance between the fake entry points and the address location is euclidean, or calculated with the network.}

\item{pseudo_entrance}{The possibility to calculate the distance to created pseudo entrances. These pseudo entrances are created,
by intersecting a buffer that was created around a greenspace with network points. default is \code{FALSE}}

\item{entrances_within_buffer}{The possibility to get the closest greenspace for every point. default is \code{FALSE}}

\item{speed}{A numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{A numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{city}{When using a network buffer, you can add a city where your address points are to speed up the process}

\item{UID}{A character string representing a unique identifier for each point of interest}
}
\value{
The distance to the nearest greenspace and whether the greenspace are within the buffer distance that is given
}
\description{
Distance to the nearest fake entries of greenspace
}
