% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_coverage.R
\name{greenspace_pct}
\alias{greenspace_pct}
\title{greenspace Coverage}
\usage{
greenspace_pct(
  address_location,
  greenspace_layer = NULL,
  buffer_distance = NULL,
  network_buffer = FALSE,
  folder_path_network = NULL,
  speed = NULL,
  time = NULL,
  epsg_code = NULL,
  UID = NULL,
  network_file = NULL,
  city = NULL,
  address_location_neighborhood = FALSE
)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{greenspace_layer}{A greenspace layer, the layer should be in projected coordinates}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{network_buffer}{A logical, the default is an euclidean buffer, when TRUE, a network buffer will be used.}

\item{folder_path_network}{optional; Folder path to where the retrieved network should be saved continuously. Must not include a filename extension (e.g. '.shp', '.gpkg').}

\item{speed}{A numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{epsg_code}{A espg code to get a Projected CRS in the final output, If missing, the default is \code{3395}}

\item{UID}{A character string representing a unique identifier for each point of interest}

\item{city}{When using a network buffer, you can add a city where your address points are to speed up the process}

\item{address_location_neighborhood}{A logical, indicating whether to calculate with an address point or a neighbourhood. default is \code{FALSE}}

\item{download_dir}{A directory to download the network file, the default will be \code{tempdir()}.}
}
\value{
Returns the percentage of greenspace coverage given a certain buffer.
}
\description{
greenspace Coverage
}
