% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_cover.R
\name{land_cover}
\alias{land_cover}
\title{Calculate the percentage of area covered by each land cover class within a given buffer distance or location}
\usage{
land_cover(
  address_location,
  raster,
  buffer_distance = NULL,
  network_buffer = FALSE,
  folder_path_network = NULL,
  folder_path_land_cover = NULL,
  network_file = NULL,
  epsg_code = NULL,
  UID = NULL,
  address_location_neighborhood = FALSE,
  speed = NULL,
  time = NULL,
  city = NULL,
  year = "2021",
  plot_landcover = FALSE
)
}
\arguments{
\item{address_location}{A \code{\link[sf]{sf}} object representing the location of interest, the location should be in projected coordinates.}

\item{raster}{Optional; a object of class \code{\link[terra]{rast}} with land cover values, if missing planetary computer will be used to extract the land cover values.}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{network_buffer}{Optional; a  logical, the default is an euclidean buffer, when TRUE, a network buffer will be used.}

\item{folder_path_network}{optional; Folder path to where the retrieved network should be saved continuously. Must not include a filename extension (e.g. '.shp', '.gpkg').}

\item{folder_path_land_cover}{optional; Folder path to where the retrieved land cover should be saved continuously. Must not include a filename extension.}

\item{network_file}{Optional; a \code{\link[sfnetworks]{sfnetwork}} object representing a road network, If missing the road network will be created.}

\item{epsg_code}{Optional; a  epsg code to get a Projected CRS in the final output, If missing, the default is \code{3395}}

\item{UID}{Optional; a  character string representing a unique identifier for each point of interest}

\item{address_location_neighborhood}{Optional; a  logical, indicating whether to calculate with an address point or a neighbourhood. default is \code{FALSE}}

\item{speed}{Optional; a  numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{Optional; a  numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{city}{Optional; when using a network buffer, you can add a city where your address points are to speed up the process}

\item{year}{Optional; The year of the satellite images. The years 2020 and 2021 can be used for the time being.}

\item{plot_landcover}{Optional; to plot the land cover, default is \code{FALSE}}
}
\value{
The percentage of each land cover type within a given buffer or isochrone around a set of locations is printed.
}
\description{
Calculate the percentage of area covered by each land cover class within a given buffer distance or location
}
\examples{
# Read a dataset, in this instance we will use the first ten neighborhoods from the
# Ams_Neighborhood dataset, which contains polygon geometry
df <- Ams_Neighborhoods[1:10,]

# calculate the land cover values for the neighborhoods with default settings
land_cover(df, buffer_distance=300)

# calculate the land cover values for neighborhoods, using the neighborhood as buffer
land_cover(df, address_location_neighborhood = TRUE)

# calculate land_cover values for the neighborhood, using the network as buffer
land_cover(df, buffer_distance=300, network_buffer=TRUE)


}
