% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_file}
\alias{network_file}
\title{an sf data frame of a small network in Amsterdam
network is represented with linestrings}
\format{
A data frame with 3899 rows and 3 variables:
\describe{
\item{osm_id}{The ID of the street}
\item{name}{The name of the street}
\item{geom}{The geometry of the LINESTRING}
}
}
\source{
{retrieved from osmextract, https://cran.r-project.org/web/packages/osmextract/vignettes/osmextract.html}
}
\usage{
network_file
}
\description{
Contains 3899 linestring of roads in Amsterdam
}
\examples{
data(network_file)     # Lazy loading

}
\keyword{datasets}
