% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_coverage.R
\name{park_pct}
\alias{park_pct}
\title{Park Coverage}
\usage{
park_pct(address_location, park_layer, buffer_distance, net, UID)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{park_layer}{A park layer, the layer should be in projected coordinates}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{net}{an optional \code{sfnetworks} object representing a road network}

\item{UID}{A character string representing a unique identifier for each point of interest}
}
\value{
Returns the percentage of park coverage given a certain buffer.
}
\description{
Park Coverage
}
