% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{parks_access}
\alias{parks_access}
\title{Calculating the distance to the nearest park access given a location.}
\usage{
parks_access(address, buffer_distance = 300, net, parks, UID)
}
\arguments{
\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{net}{An optional sfnetwork object representing a road network}

\item{parks}{A spatial object representing parks}

\item{UID}{A character string representing a unique identifier for each point of interest}

\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}
}
\value{
A sf data frame with the spatial object that was used as input, the nearest park in meters, and whether or not there is a park within the buffer distance
}
\description{
Calculating the distance to the nearest park access given a location.
}
