% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility_fake_entrance.R
\name{parks_access_entrance}
\alias{parks_access_entrance}
\title{Distance to the nearest fake entries of parks}
\usage{
parks_access_entrance(address, buffer_distance = 300, net, parks, UID)
}
\arguments{
\item{address}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{net}{An optional \code{sfnetworks} object representing a road network}

\item{parks}{A spatial object representing parks}

\item{UID}{A character string representing a unique identifier for each point of interest}
}
\value{
The distance to the nearest fake entries of parks and whether the parks are within the buffer distance that is given
}
\description{
Distance to the nearest fake entries of parks
}
