% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution_analysis.R
\name{resolution_analysis}
\alias{resolution_analysis}
\title{Viewshed Resolution Analysis}
\usage{
resolution_analysis(
  observer,
  dsm_rast,
  dtm_rast,
  greenspace_rast = NULL,
  max_distance = 800,
  observer_height = 1.7,
  raster_res = NULL,
  progress = FALSE
)
}
\arguments{
\item{observer}{object of class \code{sf} with POINT geometries; Observer location(s) from where the resolution analysis should be computed.}

\item{dsm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DSM}

\item{dtm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DTM}

\item{max_distance}{numeric; Buffer distance to calculate the viewsheds}

\item{observer_height}{numeric > 0; Height of the observer (e.g. 1.7 meters)}

\item{raster_res}{integer vector; Resolution values that the viewshed raster should be aggregated to. All values must be a multible of the dsm_rast resolution}
}
\value{
object of class \code{\link[tibble]{tibble}}. Resolution, Similarity, Time \link{ms}
}
\description{
Aggregating the input DSM reduces computiation time on cost of accuracy.
The \code{resolution_analysis} function computes viewsheds at different resolutions and compares computaion time and accuracy to the original viewshed
}
