% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_rast.R
\name{sf_to_rast}
\alias{sf_to_rast}
\title{sf to raster interpolation}
\usage{
sf_to_rast(
  observer,
  v,
  aoi = NULL,
  max_distance = Inf,
  n = Inf,
  beta = 2,
  raster_res = NULL,
  spacing = raster_res,
  na_only = FALSE,
  cores = 1,
  progress = FALSE
)
}
\arguments{
\item{observer}{object of class \code{sf}; Observer location(s) used for the interpolation. See ‘Details’ for valid sf geometry types}

\item{v}{character; Name of a variable in \code{observer} which values are used for the interpolation}

\item{aoi}{optional; object of class \code{sf}; Area of interest the raster map should me masked to. If NULL, the bounding box of \code{observer} will be used instead}

\item{max_distance}{numeric; Only observer locations within the buffer distance from the prediction location are used for interpolation}

\item{n}{numeric; Number of nearest observer locations that should be used for the interpolation}

\item{beta}{numeric; Inverse istance power, β, determines the degree to which the nearer observer locations are preferred over more distant points}

\item{raster_res}{numeric; Resolution of the interpolatet raster map that is returned}

\item{spacing}{optional; numeric; Only if \code{observer} is a linestring (or polygon), points on the line (or on a grid) will be generated.
The \code{spacing} parameter sets the distance in between the points on the line/grid}

\item{cores}{numeric; The number of cores to use, i.e. at most how many child processes will be run simultaneously}

\item{progress}{logical; Show progress bar?}
}
\value{
object of class \code{\link[terra]{rast}}
}
\description{
Function for computing a continuous raster map from a sf object by interpolating the sf values. For the interpolation Inverse Distance Weighting (IDW) is beeing used.
}
\details{
\code{observer} needs to be a geometry of type POINT, LINESTRING, MULTILINESTRING, POLYGON or MULTIPOLYGON. If \code{observer} is a LINESTRING or MULTILINESTRING,
points will be generated along the line(s) every \code{raster_res} meters. If \code{observer} is a POLYGON or MULTIPOLYGON, a grid with resolution = \code{raster_res}
will be generated as the obersver locations.
The CRS (\code{\link[sf]{st_crs}}) needs to have a metric unit!
}
