% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewshed.R
\name{viewshed}
\alias{viewshed}
\title{Viewshed}
\usage{
viewshed(
  observer,
  dsm_rast,
  dtm_rast,
  max_distance = 800,
  observer_height = 1.7,
  raster_res = NULL,
  plot = FALSE
)
}
\arguments{
\item{observer}{object of class \code{sf} with one point; Starting location}

\item{dsm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DSM}

\item{dtm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DTM}

\item{max_distance}{numeric; Buffer distance to calculate the viewshed}

\item{observer_height}{numeric > 0; Height of the observer (e.g. 1.7 meters)}

\item{raster_res}{optional; NULL or numeric > 0; Resolution that the viewshed raster should be aggregated to. Must be a multible of the dsm_rast resolution}

\item{plot}{optional; Plot the intersect of the buffer around the observer location and the DSM (left DSM; right visibility raster)}
}
\value{
object of class \code{\link[terra]{rast}}
}
\description{
Computes a binary viewshed of a point on a Digital Surface Model (DSM). The observer height is based on the heigth of a Digital Terrain Model (DTM).
}
