% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility_proportion.R
\name{visibility_proportion}
\alias{visibility_proportion}
\title{Visibility Proportion}
\usage{
visibility_proportion(
  observer,
  max_distance = 800,
  dsm_rast,
  dtm_rast,
  observer_height = 1.7,
  raster_res = NULL,
  plot = FALSE
)
}
\arguments{
\item{observer}{object of class \code{sf} with one point; Starting location}

\item{max_distance}{numeric; Buffer distance to calculate the viewshed}

\item{dsm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DSM}

\item{dtm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DTM}

\item{observer_height}{numeric > 0; Height of the observer (e.g. 1.7 meters)}

\item{raster_res}{optional; NULL or numeric > 0; Resolution that the viewshed raster should be aggregated to. Must be a multible of the dsm_rast resolution}

\item{plot}{optional; Plot the intersect of the buffer around the observer location and the DSM (left DSM; right visibility raster)}
}
\value{
numeric; Proportion of visible area of the viewshed. Values range from 0 to 1
}
\description{
Function that calculates the proportion of visible area within a viewshed
}
