% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeWeights.R
\name{visualizeWeights}
\alias{visualizeWeights}
\title{Visualize Weights}
\usage{
visualizeWeights(x, m = 0.5, b = 8, mode = c("logit", "exponential"))
}
\arguments{
\item{x}{numeric or object of class \code{\link[terra]{rast}}; Either numeric, indicating the buffer distance or a \code{\link[terra]{rast}} of the \code{\link[GreenExp]{viewshed}} function.}

\item{m}{numeric; See ‘Details’.}

\item{b}{numeric; See ‘Details’.}

\item{mode}{character; 'logit' or 'exponential'. See ‘Details’.}
}
\description{
Helper function to visualize the parameters used in the \code{\link[GreenExp]{vgvi_from_sf}} function.
}
\details{
The type of function, used for calculating the distance decay weights, can be defined with the \code{mode} parameter.
The argument 'logit' uses the logistic function, d = 1 / (1 + e^(b * (x - m))) and 'exponential' the exponential function d = 1 / (1 + (b * x^m)).
The decay function can be visualized using the \code{\link[GreenExp]{viewshed}} function.
}
