% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_calc.R
\name{calc_ndvi}
\alias{calc_ndvi}
\title{Creating average NDVI values per location}
\usage{
calc_ndvi(
  address_location,
  raster,
  buffer_distance = NULL,
  network_buffer = FALSE,
  download_dir = tempdir(),
  epsg_code = NULL,
  network_file = NULL,
  UID = NULL,
  address_location_neighborhood = FALSE,
  speed = NULL,
  time = NULL,
  engine = "pc",
  city = NULL,
  start_date = "2020-01-01",
  end_date = "2021-01-01",
  save_NDVI = FALSE,
  plot_NDVI = FALSE
)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{raster}{raster file with NDVI values}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{network_buffer}{A logical, the default is an euclidean buffer, when TRUE, a network buffer will be used.}

\item{download_dir}{A directory to download the network file, the default will be \code{tempdir()}.}

\item{epsg_code}{A espg code to get a Projected CRS in the final output, If missing, the default is \code{3395}}

\item{network_file}{An optional sfnetwork object representing a road network, If missing the road network will be created.}

\item{UID}{A character string representing a unique identifier for each point of interest}

\item{address_location_neighborhood}{A logical, indicating whether to calculate with an address point or a neighbourhood. default is \code{FALSE}}

\item{speed}{A numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{A numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{engine}{When the raster is missing, users can choose whether they want to use Google Earth engine \code{gee} or Planetary Computer \code{pc} to calculate the ndvi}

\item{city}{When using a network buffer, you can add a city where your address points are to speed up the process}

\item{start_date}{The start date from when the satellite images will be filtered \code{yyyy-mm-dd} default = \code{2020-01-01}}

\item{end_date}{The end date from when the satellite images will be filtered. \code{yyyy-mm-dd} default = \code{2021-01-01}}

\item{save_NDVI}{If you want to save the NDVI values, default is \code{FALSE}}

\item{plot_NDVI}{If you want to plot the NDVI, default is \code{FALSE}}
}
\value{
A \code{sf} dataframe with the mean ndvi, the geometry and the buffer that was used
}
\description{
Creating average NDVI values per location
}
