% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_cover.R
\name{land_cover}
\alias{land_cover}
\title{Calculate the percentage of area covered by each land cover class within a given buffer distance or location}
\usage{
land_cover(
  address_location,
  class_raster,
  buffer_distance,
  net,
  UID,
  address_calculation = TRUE,
  speed,
  time
)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{class_raster}{A raster object representing a land cover classification map, where each pixel is assigned a land cover class}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{net}{an optional sfnetwork object representing a road network}

\item{UID}{A character string representing a unique identifier for each point of interest}

\item{address_calculation}{A logical, indicating whether to calculate the address location (if not a point) as the centroid of the polygon containing it (default is 'TRUE')}

\item{speed}{A numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{A numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}
}
\value{
The percentage of each land cover type within a given buffer or isochrone around a set of locations is printed.
}
\description{
Calculate the percentage of area covered by each land cover class within a given buffer distance or location
}
