% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_cover.R
\name{land_cover}
\alias{land_cover}
\title{Calculate the percentage of area covered by each land cover class within a given buffer distance or location}
\usage{
land_cover(
  address_location,
  raster,
  buffer_distance = NULL,
  network_buffer = FALSE,
  download_dir = tempdir(),
  network_file = NULL,
  epsg_code = NULL,
  UID = NULL,
  address_calculation = TRUE,
  speed = NULL,
  time = NULL,
  city = NULL,
  year = "2021",
  plot_landcover = FALSE
)
}
\arguments{
\item{address_location}{A spatial object representing the location of interest, the location should be in projected coordinates.}

\item{raster}{raster file with land cover values, if raster file is missing planetary computer will be used.}

\item{buffer_distance}{A distance in meters to create a buffer or isochrone around the address location}

\item{network_buffer}{A logical, the default is an euclidean buffer, when TRUE, a network buffer will be used.}

\item{download_dir}{A directory to download the network file, the default will be \code{tempdir()}.}

\item{network_file}{n optional sfnetwork object representing a road network, If missing the road network will be created.}

\item{epsg_code}{A espg code to get a Projected CRS in the final output, If missing, the default is \code{3395}}

\item{UID}{A character string representing a unique identifier for each point of interest}

\item{address_calculation}{A logical, indicating whether to calculate the address location (if not a point) as the centroid of the polygon containing it (default is 'TRUE')}

\item{speed}{A numeric value representing the speed in km/h to calculate the buffer distance (required if \code{time} is provided)}

\item{time}{A numeric value representing the travel time in minutes to calculate the buffer distance (required if \code{speed} is provided)}

\item{city}{When using a network buffer, you can add a city where your address points are to speed up the process}

\item{year}{The year of the satellite images. The years 2020 and 2021 can be used for the time being.}

\item{plot_landcover}{Option to plot the land cover, default is \code{FALSE}}
}
\value{
The percentage of each land cover type within a given buffer or isochrone around a set of locations is printed.
}
\description{
Calculate the percentage of area covered by each land cover class within a given buffer distance or location
}
